feat(AddToCartTreatments.tsx): add useState hook for disabled state feat(AddToCartTreatments.tsx): add useFetchers hook fix(Layout.tsx): conditionally render components based on path feat(ArtistHero.tsx): add Box and Button components fix(ArtistServiceContent.tsx): add sizes prop to Image component fix(($locale).artist_.$username.treatment.$productHandle.tsx): change mih value to dynamic calculation fix(($locale).products.$handle.tsx): export redirectToFirstVariant function fix(AddToCartButton.tsx): add onFinish prop delete(($locale).treatments.$productHandle._index.tsx): delete unused file delete(($locale).treatments.$productHandle.completed.tsx): delete unused file
chore(routes): remove unused pick-datetime, pick-location, and pick-more routes

chore(pick-username.tsx): remove unused imports and code

fix(routes): change username parameter to not be included in shouldRevalidate function
feat(routes): add support for picking artists and updating URL with selected artist
